requirements.txt
mcp>=1.0.0
openai>=1.50.0
sqlite-vec>=0.1.6
pydantic>=2.0.0
tenacity>=8.2.0
streamlit>=1.30.0
dateparser>=1.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
click>=8.1.0
httpx>=0.27.0
oss2>=2.18.0
PyMuPDF>=1.23.0
